Code improvement: access to the bank only when necessary
